The iTunes image is now also used as the RSS image if no RSS image is specified.
Warning message added to podcast channel feeds that are password protected. Warning: This feed is password protected, it cannot be accessed by public services such as feedvalidator.org or the iTunes podcast directory. (we're getting a lot of support questions about this recently. If your feed is is password protected, then services and directories can't access it by design.)


git-svn-id: https://plugins.svn.wordpress.org/powerpress/trunk@583296 b8457f37-d9ea-0310-8a92-e5e31aec5664